+++ src/main/scala/enumeration/ProbCosts.scala
@@ -3,9 +3,9 @@ package enumeration
 import ast.ASTNode
 
 object ProbCosts{
-  def getCosts(childrenCost: Double, candidateChildren: List[ASTNode], childrenArity: Int): List[List[Double]] = {
+  def getCosts(childrenCost: Double, candidateChildren: List[ASTNode], childrenArity: Int): List[List[Int]] = {
     val childrenCosts = candidateChildren.map(c => c.cost)
-    var candidateCosts = List[List[Double]]()
+    var candidateCosts = List[List[Int]]()
 
     val combinations = childrenCosts.combinations(childrenArity).toList.filter(c => c.sum == childrenCost)
     candidateCosts = combinations.map(c => c.permutations.toList).flatten
